Refactored data_file.py into the crds.io package,  preserving the data_file module as an interface for now.  This sets the stage for expanding polymorphic features of the various I/O classes,  shortly to sprout new support for file format info and array data access to enable more advanced certifier checks on data.
