+++ crds/core/python23.py
@@ -9,6 +9,7 @@ if sys.version_info >= (3,0,0):
     from html import unescape
     import configparser
     import pickle
+    from io import StringIO
 
     def unicode_to_str(input):
         """Recursively convert .json inputs with unicode to simple Python strings."""
@@ -23,6 +24,7 @@ else:
 
     import ConfigParser as configparser
     import cPickle as pickle
+    import StringIO
 
     def unicode_to_str(input):
         """Recursively convert .json inputs with unicode to simple Python strings."""
